FEATURE: Hide webp files in file list
When browsers request jpg or png files and a corresponding webp file
is available the webp file is returned instead. Webp files are created
automatically.

This patch now hides the automatically created webp files in the TYPO3
because having those in place is an implementation detail when it comes
to editors selecting files.